Added Hibernate schema validation to the JPA example and refined JPA models where necessary

+++ examples/jpa-sqlite-example/src/main/java/com/github/appreciated/vortex_crud/example/jpa/entity/TaskHasTask.java
@@ -1,7 +1,6 @@
 package com.github.appreciated.vortex_crud.example.jpa.entity;
 
 import com.github.appreciated.vortex_crud.core.ui.factories.form.elements.fields.functions.DateTimePickerFactory;
-import com.github.appreciated.vortex_crud.core.ui.factories.form.elements.fields.functions.TextFieldFactory;
 import com.github.appreciated.vortex_crud.jpa.service.JpaFieldRenderer;
 import jakarta.persistence.*;
 
@@ -13,7 +12,7 @@ public class TaskHasTask {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "task_id", nullable = false)
@@ -27,11 +26,11 @@ public class TaskHasTask {
     private LocalDateTime createdAt;
 
     // Getters and Setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
